fix: move location of preview files to design-tokens
