+++ packages/design-tokens/design-token-previews/borders.mdx
@@ -1,6 +1,6 @@
-import BorderStyleTokensTable from "../../../.storybook/blocks/BorderStyleTokensTable";
-import BorderWidthTokensTable from "../../../.storybook/blocks/BorderWidthTokensTable";
-import BorderRadiiTokensTable from "../../../.storybook/blocks/BorderRadiiTokensTable";
+import { BorderStyleTokensTable } from "./tables/BorderStyleTokensTable";
+import { BorderWidthTokensTable } from "./tables/BorderWidthTokensTable";
+import { BorderRadiiTokensTable } from "./tables/BorderRadiiTokensTable";
 
 # Border Tokens
 
@@ -14,6 +14,8 @@ import BorderRadiiTokensTable from "../../../.storybook/blocks/BorderRadiiTokens
 
 ## Border Widths
 
-<BorderWidthTokensTable /># ## Border Radii{" "}
+<BorderWidthTokensTable />
+
+## Border Radii
 
 <BorderRadiiTokensTable />
