+++ packages/design-tokens/design-token-previews/colors.mdx
@@ -1,4 +1,4 @@
-import ColorTokensTable from "../../../.storybook/blocks/ColorTokensTable";
+import { ColorTokensTable } from "./tables/ColorTokensTable";
 
 # Color Tokens
