+++ packages/design-tokens/design-token-previews/font-sizes.mdx
@@ -1,4 +1,4 @@
-import FontSizeTokensTable from "../../../.storybook/blocks/FontSizeTokensTable";
+import { FontSizeTokensTable } from "./tables/FontSizeTokensTable";
 
 # Font Size Tokens
