+++ packages/design-tokens/design-token-previews/font-weights.mdx
@@ -1,4 +1,4 @@
-import FontWeightTokensTable from "../../../.storybook/blocks/FontWeightTokensTable";
+import { FontWeightTokensTable } from "./tables/FontWeightTokensTable";
 
 # Font Weight Tokens
