+++ packages/design-tokens/design-token-previews/icon-sizes.mdx
@@ -1,4 +1,4 @@
-import IconSizeTokensTable from "../../../.storybook/blocks/IconSizeTokensTable";
+import { IconSizeTokensTable } from "./tables/IconSizeTokensTable";
 
 # Icon Size Tokens
