+++ packages/design-tokens/design-token-previews/spacing.mdx
@@ -1,4 +1,4 @@
-import SpacingTokensTable from "../../../.storybook/blocks/SpacingTokensTable";
+import { SpacingTokensTable } from "./tables/SpacingTokensTable";
 
 # Spacing Tokens
