+++ packages/design-tokens/design-token-previews/tables/FontSizeTokensTable.tsx
@@ -1,8 +1,9 @@
 import React from "react";
-import spaceTokens from "../../packages/design-tokens/src/tokens/space.tokens.json";
+import map from "lodash/map";
+import fontSizeTokens from "../../src/tokens/font-size.tokens.json";
 import { formatTokenName, formatRems, formatPixels } from "./utils";
 
-const SpacingTokensTable = () => {
+export const FontSizeTokensTable = (): JSX.Element => {
   return (
     <table style={{ width: "100%" }}>
       <thead>
@@ -22,23 +23,18 @@ const SpacingTokensTable = () => {
         </tr>
       </thead>
       <tbody>
-        {Object.entries(spaceTokens.space).map((token) => {
+        {map(Object.entries(fontSizeTokens["font-size"]), (token) => {
           return (
             <tr key={token[0]}>
-              <td>{formatTokenName(token, "space")}</td>
+              <td>{formatTokenName(token, "fontWeight")}</td>
               <td>{formatPixels(token)}</td>
               <td>{formatRems(token)}</td>
               <td>
-                {!token[1].value.includes("-") && (
-                  <div
-                    style={{
-                      width: `${formatPixels(token)}`,
-                      backgroundColor: "#413cff",
-                    }}
-                  >
-                    &nbsp;
-                  </div>
-                )}
+                <div style={{ paddingTop: "1rem", paddingBottom: "1rem" }}>
+                  <text style={{ fontSize: token[1].value }}>
+                    The quick brown fox jumped over the lazy dog.
+                  </text>
+                </div>
               </td>
             </tr>
           );
@@ -47,5 +43,3 @@ const SpacingTokensTable = () => {
     </table>
   );
 };
-
-export default SpacingTokensTable;
